+++ src/main/java/soma/ghostrunner/domain/auth/application/AuthService.java
@@ -12,7 +12,7 @@ import soma.ghostrunner.domain.auth.resolver.AuthIdResolver;
 import soma.ghostrunner.domain.auth.api.dto.request.SignUpRequest;
 import soma.ghostrunner.domain.auth.api.dto.TermsAgreementDto;
 import soma.ghostrunner.domain.member.application.MemberService;
-import soma.ghostrunner.domain.member.Member;
+import soma.ghostrunner.domain.member.domain.Member;
 import soma.ghostrunner.domain.member.application.dto.MemberMapper;
 import soma.ghostrunner.domain.member.domain.TermsAgreement;
 import soma.ghostrunner.global.error.ErrorCode;
