+++ src/main/java/soma/ghostrunner/domain/member/MemberRepository.java
@@ -3,6 +3,7 @@ package soma.ghostrunner.domain.member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+import soma.ghostrunner.domain.member.domain.Member;
 
 import java.util.Optional;
