+++ src/main/java/soma/ghostrunner/domain/member/domain/Member.java
@@ -1,7 +1,8 @@
-package soma.ghostrunner.domain.member;
+package soma.ghostrunner.domain.member.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
+import soma.ghostrunner.domain.member.MemberBioInfo;
 import soma.ghostrunner.global.common.BaseTimeEntity;
 import soma.ghostrunner.global.common.document.TestOnly;
