+++ src/test/java/soma/ghostrunner/domain/member/MemberRepositoryTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import soma.ghostrunner.IntegrationTestSupport;
 import soma.ghostrunner.domain.member.dao.MemberAuthInfoRepository;
+import soma.ghostrunner.domain.member.domain.Member;
 import soma.ghostrunner.domain.member.domain.MemberAuthInfo;
 
 class MemberRepositoryTest extends IntegrationTestSupport {
